# glasses-store/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Optical E-commerce API System
  description: |
    API Contract for Glass Store System.
    Architecture designed for scalability from Frame-only MVP to full Optical Lab integration.
    
    Versioning Strategy:
    - /api/v1: Frame Only (MVP Production)
    - /api/v2: Optical Extension (Lens + Prescription + Lab) - Reserved
    - /api/v3: AI + AR Try On - Reserved
  version: 1.0.0
  contact:
    name: Backend Architect Team
servers:
  - url: https://api.glass-store.com
    description: Production Server
  - url: http://localhost:8080
    description: Local Development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          $ref: '#/components/responses/SuccessResponse'
  /api/v1/auth/login:
    post:
      tags: [Auth]
      security: []
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TokenResponse'
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
  /api/v1/auth/forgot-password:
    post:
      tags: [Auth]
      security: []
      summary: Request password reset
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
  /api/v1/auth/reset-password:
    post:
      tags: [Auth]
      security: []
      summary: Reset password with token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
                newPassword: { type: string }
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      security:
        - bearerAuth: []
      summary: Logout user
      description: Revoke refresh token and invalidate session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
  /api/v1/auth/change-password:
    post:
      tags: [Auth]
      security:
        - bearerAuth: []
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
  /api/v1/auth/verify-email:
    post:
      tags: [Auth]
      security: []
      summary: Verify email with code
      parameters:
        - name: token
          in: query
          description: Verification token sent via email
          schema: { type: string }
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
  /api/v1/auth/resend-verification:
    post:
      tags: [Auth]
      security: []
      summary: Resend verification email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'

  # --- USER ---
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    SuccessResponse:
      description: Successful operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
  schemas:
...
    ApiResponse:
      type: object
      properties:
        status:
          type: integer
          example: 200
        message:
          type: string
          example: "Success"
        data:
          type: object
          nullable: true
        errors:
          type: array
          nullable: true
          items:
            type: object
            properties:
              code: { type: string }
              detail: { type: string }

    # --- ENUMS ---
...
    RegisterRequest:
      type: object
      required: [email, password, fullName]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        fullName: { type: string }
        phone: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    TokenResponse:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        tokenType: { type: string, example: "Bearer" }
        expiresIn: { type: integer, example: 3600 }
    ChangePasswordRequest:
      type: object
      required: [oldPassword, newPassword, confirmPassword]
      properties:
        oldPassword: { type: string }
        newPassword: { type: string, minLength: 8 }
        confirmPassword: { type: string }
...